Combine check in and check in by barcode resources CIRC-154
